# internal/openapi/openapi.json
{
  "components": {
    "responses": {
...
      "InvalidAPIKey": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "invalid_api_key",
                "message": "Provide an active project environment API key."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The project API key is missing, malformed, revoked, replaced, or inactive."
      },
...
      "MeterConflict": {
        "content": {
          "application/json": {
            "examples": {
              "meter_already_exists": {
                "value": {
                  "error": {
                    "code": "meter_already_exists"
                  }
                }
              },
              "meter_definition_conflict": {
                "value": {
                  "error": {
                    "code": "meter_definition_conflict"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter cannot be created in the API key's environment."
      },
      "MeterCreateConflict": {
        "content": {
          "application/json": {
            "examples": {
              "environment_inactive": {
                "value": {
                  "error": {
                    "code": "environment_inactive"
                  }
                }
              },
              "meter_already_exists": {
                "value": {
                  "error": {
                    "code": "meter_already_exists"
                  }
                }
              },
              "meter_definition_conflict": {
                "value": {
                  "error": {
                    "code": "meter_definition_conflict"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter cannot be created in the selected environment."
      },
      "MeterEnvironmentConflict": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "environment_inactive",
                "message": "Activate Live before managing its meters."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The selected project environment is not active."
      },
      "MeterEnvironmentNotFound": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "project_environment_not_found",
                "message": "This project environment is not available."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The selected project environment is unavailable."
      },
      "MeterFailed": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "meter_operation_failed",
                "message": "Consumel could not complete the meter request. Try again shortly."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter request could not be completed."
      },
      "MeterInvalid": {
        "content": {
          "application/json": {
            "examples": {
              "invalid_meter": {
                "value": {
                  "error": {
                    "code": "invalid_meter"
                  }
                }
              },
              "invalid_request": {
                "value": {
                  "error": {
                    "code": "invalid_request"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter request or key is invalid."
      },
      "MeterListInvalid": {
        "content": {
          "application/json": {
            "examples": {
              "invalid_cursor": {
                "value": {
                  "error": {
                    "code": "invalid_cursor"
                  }
                }
              },
              "invalid_request": {
                "value": {
                  "error": {
                    "code": "invalid_request"
                  }
                }
              },
              "invalid_search": {
                "value": {
                  "error": {
                    "code": "invalid_search"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The list pagination or search input is invalid."
      },
      "MeterNotFound": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "meter_not_found",
                "message": "This meter does not exist in the selected environment."
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter is not present in the authenticated environment."
      },
      "MeterOrEnvironmentNotFound": {
        "content": {
          "application/json": {
            "examples": {
              "meter_not_found": {
                "value": {
                  "error": {
                    "code": "meter_not_found"
                  }
                }
              },
              "project_environment_not_found": {
                "value": {
                  "error": {
                    "code": "project_environment_not_found"
                  }
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The meter or selected project environment is unavailable."
      },
      "NotAuthenticated": {
        "content": {
          "application/json": {
            "example": {
              "error": {
                "code": "not_authenticated"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authentication is required."
      },
...
    },
    "schemas": {
...
      "CreateMeterRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "example": "Requests processed by your API.",
            "maxLength": 500,
            "minLength": 1,
            "type": [
              "string",
              "null"
            ]
          },
          "meterKey": {
            "example": "api_calls",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "type": "string"
          },
          "name": {
            "example": "API calls",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "enum": [
              "prepaid",
              "postpaid",
              "hybrid"
            ],
            "example": "postpaid",
            "type": "string"
          }
        },
        "required": [
          "meterKey",
          "name",
          "type"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Meter": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "meterKey": {
            "maxLength": 120,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "type": "string"
          },
          "type": {
            "enum": [
              "prepaid",
              "postpaid",
              "hybrid"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "meterKey",
          "name",
          "description",
          "type",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "Meters": {
        "additionalProperties": false,
        "properties": {
          "meters": {
            "items": {
              "$ref": "#/components/schemas/Meter"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "meters",
          "nextCursor"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "localCookieSession": {
        "in": "cookie",
        "name": "consumel_session",
        "type": "apiKey"
      },
      "productionCookieSession": {
        "in": "cookie",
        "name": "__Host-consumel_session",
        "type": "apiKey"
      },
      "projectAPIKey": {
        "bearerFormat": "cm_test_… or cm_live_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Consumel usage-based billing infrastructure API.",
    "title": "Consumel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/meters": {
      "get": {
        "parameters": [
          {
            "description": "A case-insensitive meter key or name search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of meters to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "meters": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "description": "Requests processed by your API.",
                      "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                      "meterKey": "api_calls",
                      "name": "API calls",
                      "type": "postpaid",
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/Meters"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/MeterListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "List meters in the API key's project environment.",
        "tags": [
          "Meters"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "description": "Requests processed by your API.",
                "meterKey": "api_calls",
                "name": "API calls",
                "type": "postpaid"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateMeterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "description": "Requests processed by your API.",
                  "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                  "meterKey": "api_calls",
                  "name": "API calls",
                  "type": "postpaid",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "$ref": "#/components/responses/MeterInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "409": {
            "$ref": "#/components/responses/MeterConflict"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Create a meter in the API key's project environment.",
        "tags": [
          "Meters"
        ]
      }
    },
    "/v1/meters/{meterKey}": {
      "get": {
        "parameters": [
          {
            "description": "The stable meter key supplied when the meter was created.",
            "in": "path",
            "name": "meterKey",
            "required": true,
            "schema": {
              "maxLength": 120,
              "pattern": "^[a-z][a-z0-9_-]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "description": "Requests processed by your API.",
                  "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                  "meterKey": "api_calls",
                  "name": "API calls",
                  "type": "postpaid",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/MeterInvalid"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAPIKey"
          },
          "404": {
            "$ref": "#/components/responses/MeterNotFound"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "projectAPIKey": []
          }
        ],
        "summary": "Return one meter from the API key's project environment.",
        "tags": [
          "Meters"
        ]
      }
    },
...
    "/v1/projects/{projectId}/environments/{environment}/meters": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "A case-insensitive meter key or name search.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "description": "The opaque next cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of meters to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "meters": [
                    {
                      "createdAt": "2026-09-28T12:00:00Z",
                      "description": "Requests processed by your API.",
                      "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                      "meterKey": "api_calls",
                      "name": "API calls",
                      "type": "postpaid",
                      "updatedAt": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "nextCursor": null
                },
                "schema": {
                  "$ref": "#/components/schemas/Meters"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/MeterListInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/MeterEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/MeterEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "List meters in the signed-in project workspace.",
        "tags": [
          "Dashboard Meters"
        ]
      },
      "post": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "description": "Requests processed by your API.",
                "meterKey": "api_calls",
                "name": "API calls",
                "type": "postpaid"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateMeterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "description": "Requests processed by your API.",
                  "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                  "meterKey": "api_calls",
                  "name": "API calls",
                  "type": "postpaid",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "$ref": "#/components/responses/MeterInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/MeterEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/MeterCreateConflict"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Create a meter from the signed-in project workspace.",
        "tags": [
          "Dashboard Meters"
        ]
      }
    },
    "/v1/projects/{projectId}/environments/{environment}/meters/{meterKey}": {
      "get": {
        "parameters": [
          {
            "description": "The immutable ID of the project selected in the dashboard.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The selected isolated project environment.",
            "in": "path",
            "name": "environment",
            "required": true,
            "schema": {
              "enum": [
                "sandbox",
                "live"
              ],
              "type": "string"
            }
          },
          {
            "description": "The stable meter key supplied when the meter was created.",
            "in": "path",
            "name": "meterKey",
            "required": true,
            "schema": {
              "maxLength": 120,
              "pattern": "^[a-z][a-z0-9_-]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2026-09-28T12:00:00Z",
                  "description": "Requests processed by your API.",
                  "id": "0199a9f8-f0c4-7f10-90f8-6483353e4624",
                  "meterKey": "api_calls",
                  "name": "API calls",
                  "type": "postpaid",
                  "updatedAt": "2026-09-28T12:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/Meter"
                }
              }
            },
            "description": "Completed successfully."
          },
          "400": {
            "$ref": "#/components/responses/MeterInvalid"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          },
          "404": {
            "$ref": "#/components/responses/MeterOrEnvironmentNotFound"
          },
          "409": {
            "$ref": "#/components/responses/MeterEnvironmentConflict"
          },
          "500": {
            "$ref": "#/components/responses/MeterFailed"
          }
        },
        "security": [
          {
            "productionCookieSession": []
          },
          {
            "localCookieSession": []
          }
        ],
        "summary": "Return one meter from the signed-in project workspace.",
        "tags": [
          "Dashboard Meters"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.consumel.com"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8080"
    }
  ]
}